Separate message declaration and pack/unpack implementation
